# Remove redundant CPU decomposition when done also on GPU