Factor out the initialization code needed before doing a ILU decomposition to remove the redundant software decomposition when we also perform it on the GPU. This was previously done in #4558, but lost in the adaptation to ISTLSolverEbosBda.