@blattms 2023-10-26
Not sure whether this code is still there/functional and my memory even reflects reality.
What happens for CPR on the GPU. Didn't that somehow use the AMG hierarchy from the CPU? With this it won't be there anymore.