@razvnane 2023-10-26
It is functional and it works similarly for CPR as well. I think there was a misunderstanding about what was used from CPU, but what is reused are the DUNE classes and functions regarding the creation of the CPR. These are called from the create_preconditioner_amg() function from opencl/CPR.cpp and do not depend on ISTLSolverEbos::prepare() function being called beforehand. I also tested it timing the first step of running NORNE and it shows what we would expect, i.e., a reduction in the solve time by removing the time for the preparation on the CPU (note: in the timing results the _prepare_ time is included in the _solve_ time):
```
<with removal>

CPR create_preconditioner bilu0(): 0.0608781 s
CPR create_preconditioner_amg(): 0.141993 s 
openclSolver::create_preconditioner(): 0.202939 s
openclSolver::copy_system_to_gpu(): 0.00479971 s 
openclSolver::prec_apply:  0.0482614 s
openclSolver::total_solve: 0.0539096 s
openclSolver::solve_system(): 0.0540114 s
openclSolver::get_result(): 0.000339837 s 
BlackoilModelEbos::solveJacobianSystem cum prepare time: 0.000239601
BlackoilModelEbos::solveJacobianSystem cum solve time: 0.27242
```
```
<without removal>

CPR create_preconditioner bilu0(): 0.0465751 s
CPR create_preconditioner_amg(): 0.133502 s
openclSolver::create_preconditioner(): 0.180129 s
openclSolver::copy_system_to_gpu(): 0.00479273 s
openclSolver::prec_apply:  0.0475754 s
openclSolver::total_solve: 0.0576826 s
openclSolver::solve_system(): 0.0577893 s
openclSolver::get_result(): 0.00031434 s
BlackoilModelEbos::solveJacobianSystem prepare time: 0.294071
BlackoilModelEbos::solveJacobianSystem solve time: 0.546205
```