@atgeirr 2023-10-26
jenkins build this please